// src/calc-value.enum.ts
export enum ECalcValue {
    ONE = '1',
    TWO = '2',
    THREE = '3',
    FOUR = '4',
    FIVE = '5',
    SIX = '6',
    SEVEN = '7',
    EIGHT = '8',
    NINE = '9',
    ZERO = '0',
    ADD = '+',
    SUB = '-',
    DIV = '/',
    EQUAL = '=',
    RESET = 'R',
    DELETE = 'D',
    DECIMAL_SEPARATOR = '.',
    MULTIPLY = '*',
}